#6133 - Support knowledge bases that use multiple datasets / graphs
- Extract FreeTextMultiSelect base class for Kendo multi-select with free-text support
- Reuse Validators.IRI_VALIDATOR in QuerySettingsPanel and RemoteRepositorySettingsPanel instead of hand-rolling validation
- Move datasets field from RemoteRepositorySettingsPanel to QuerySettingsPanel so both local and remote KBs can configure named-graph restrictions
- Add RemoteRepositoryUrlChangedEvent for cross-panel clearing of stale discovered datasets when endpoint URL changes
- Guard Kendo change handler binding in FreeTextMultiSelect to prevent duplicate handlers on Ajax re-render
- Refactor GeneralSettingsPanel to use FreeTextMultiSelect for additional languages
- Remove outdated documentation note claiming remote KB URL changes require application restart (they don't; syncSparqlUrlLiveParameters pushes updates live)
