+++ inception/inception-ui-kb/src/main/java/de/tudarmstadt/ukp/inception/ui/kb/project/GeneralSettingsPanel.java
@@ -23,9 +23,6 @@ import static org.apache.commons.lang3.StringUtils.isNotBlank;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.apache.commons.lang3.StringUtils;
-import org.apache.wicket.markup.head.IHeaderResponse;
-import org.apache.wicket.markup.head.OnDomReadyHeaderItem;
 import org.apache.wicket.markup.html.form.CheckBox;
 import org.apache.wicket.markup.html.form.RequiredTextField;
 import org.apache.wicket.markup.html.form.TextField;
@@ -36,16 +33,13 @@ import org.apache.wicket.spring.injection.annot.SpringBean;
 import org.apache.wicket.validation.IValidatable;
 import org.apache.wicket.validation.IValidator;
 import org.apache.wicket.validation.ValidationError;
-import org.wicketstuff.jquery.core.JQueryBehavior;
-import org.wicketstuff.jquery.core.Options;
-import org.wicketstuff.kendo.ui.KendoDataSource;
 import org.wicketstuff.kendo.ui.form.combobox.ComboBox;
-import org.wicketstuff.kendo.ui.form.multiselect.lazy.MultiSelect;
 
 import de.tudarmstadt.ukp.clarin.webanno.model.Project;
 import de.tudarmstadt.ukp.inception.kb.IriConstants;
 import de.tudarmstadt.ukp.inception.kb.KnowledgeBaseService;
 import de.tudarmstadt.ukp.inception.kb.config.KnowledgeBaseProperties;
+import de.tudarmstadt.ukp.inception.support.kendo.FreeTextMultiSelect;
 import de.tudarmstadt.ukp.inception.support.lambda.LambdaAjaxFormComponentUpdatingBehavior;
 
 public class GeneralSettingsPanel
@@ -88,23 +82,8 @@ public class GeneralSettingsPanel
         add(basePrefixField("basePrefix", "kb.basePrefix"));
         add(createCheckbox("enabled", "kb.enabled"));
 
-        var additionalLanguages = new MultiSelect<String>("additionalLanguages")
+        var additionalLanguages = new FreeTextMultiSelect("additionalLanguages")
         {
-            @Override
-            protected void onConfigure(KendoDataSource aDataSource)
-            {
-                // This ensures that we get the user input in getChoices
-                aDataSource.set("serverFiltering", true);
-            }
-
-            @Override
-            public void onConfigure(JQueryBehavior aBehavior)
-            {
-                super.onConfigure(aBehavior);
-                aBehavior.setOption("filter", Options.asString("contains"));
-                aBehavior.setOption("autoClose", false);
-            }
-
             private static final long serialVersionUID = -7735027268669019571L;
 
             @Override
@@ -119,37 +98,7 @@ public class GeneralSettingsPanel
                 }
                 return choices;
             }
-
-            @Override
-            public void convertInput()
-            {
-                var input = getInputAsArray();
-                var list = new ArrayList<String>();
-                if (input != null) {
-                    list.addAll(asList(input));
-                    list.removeIf(StringUtils::isBlank);
-                }
-                this.setConvertedInput(list);
-            }
-
-            @Override
-            public void renderHead(IHeaderResponse response)
-            {
-                super.renderHead(response);
-                // The Kendo MultiSelect with serverFiltering=true does not reliably sync chip
-                // removals back to the underlying <select> element, so removed languages still
-                // get submitted with the form. Rebuild the <select> from the widget value on
-                // every change to keep form submission in sync.
-                var script = "(function(){var attach=function(){var ms=$('#" + getMarkupId()
-                        + "').data('kendoMultiSelect');if(!ms){setTimeout(attach,50);return;}"
-                        + "ms.bind('change',function(){var v=this.value();var s=$(this.element);"
-                        + "s.empty();for(var i=0;i<v.length;i++){"
-                        + "s.append($('<option selected></option>').val(v[i]).text(v[i]));}});};"
-                        + "attach();})();";
-                response.render(OnDomReadyHeaderItem.forScript(script));
-            }
         };
-        additionalLanguages.setOutputMarkupId(true);
         additionalLanguages.setModel(kbModel.bind("kb.additionalLanguages"));
         add(additionalLanguages);
     }
