+++ inception/inception-ui-kb/src/main/java/de/tudarmstadt/ukp/inception/ui/kb/project/QuerySettingsPanel.java
@@ -18,10 +18,20 @@
 package de.tudarmstadt.ukp.inception.ui.kb.project;
 
 import static de.tudarmstadt.ukp.inception.kb.IriConstants.getFtsBackendName;
+import static java.util.Collections.emptyList;
 import static java.util.Comparator.comparing;
+import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.LinkedHashSet;
+import java.util.List;
+
+import org.apache.wicket.AttributeModifier;
 import org.apache.wicket.ajax.AjaxRequestTarget;
 import org.apache.wicket.ajax.markup.html.form.AjaxCheckBox;
+import org.apache.wicket.event.IEvent;
+import org.apache.wicket.feedback.IFeedback;
 import org.apache.wicket.markup.html.form.CheckBox;
 import org.apache.wicket.markup.html.form.DropDownChoice;
 import org.apache.wicket.markup.html.form.LambdaChoiceRenderer;
@@ -30,15 +40,23 @@ import org.apache.wicket.markup.html.form.TextField;
 import org.apache.wicket.markup.html.panel.Panel;
 import org.apache.wicket.model.CompoundPropertyModel;
 import org.apache.wicket.model.IModel;
+import org.apache.wicket.model.LambdaModel;
 import org.apache.wicket.model.Model;
+import org.apache.wicket.model.StringResourceModel;
 import org.apache.wicket.spring.injection.annot.SpringBean;
+import org.apache.wicket.validation.IValidatable;
+import org.apache.wicket.validation.Validatable;
 import org.eclipse.rdf4j.model.IRI;
+import org.wicketstuff.kendo.ui.form.multiselect.lazy.MultiSelect;
 
 import de.tudarmstadt.ukp.inception.kb.IriConstants;
 import de.tudarmstadt.ukp.inception.kb.KnowledgeBaseService;
 import de.tudarmstadt.ukp.inception.kb.config.KnowledgeBaseProperties;
 import de.tudarmstadt.ukp.inception.kb.config.KnowledgeBasePropertiesImpl;
+import de.tudarmstadt.ukp.inception.support.kendo.FreeTextMultiSelect;
+import de.tudarmstadt.ukp.inception.support.lambda.LambdaAjaxLink;
 import de.tudarmstadt.ukp.inception.support.lambda.LambdaBehavior;
+import de.tudarmstadt.ukp.inception.ui.kb.project.validators.Validators;
 
 public class QuerySettingsPanel
     extends Panel
@@ -51,7 +69,12 @@ public class QuerySettingsPanel
     private final TextField<Integer> queryLimitField;
     private final CheckBox maxQueryLimitCheckBox;
     private final CheckBox useFuzzyCheckBox;
+    private final MultiSelect<String> datasetsField;
     private final CompoundPropertyModel<KnowledgeBaseWrapper> kbModel;
+
+    // Named graphs fetched on-demand from the endpoint to offer as choices in addition to any IRIs
+    // the user enters manually.
+    private final List<String> availableDatasets = new ArrayList<>();
 
     public QuerySettingsPanel(String id, CompoundPropertyModel<KnowledgeBaseWrapper> aModel)
     {
@@ -68,6 +91,109 @@ public class QuerySettingsPanel
         maxQueryLimitCheckBox = maxQueryLimitCheckbox("maxQueryLimit", Model.of(false));
         add(maxQueryLimitCheckBox);
         add(ftsField("fullTextSearchIri", "kb.fullTextSearchIri"));
+
+        datasetsField = new FreeTextMultiSelect("datasets")
+        {
+            private static final long serialVersionUID = 1L;
+
+            @Override
+            protected List<String> getChoices(String aInput)
+            {
+                // There is no fixed vocabulary of dataset IRIs. We offer the graphs fetched from
+                // the endpoint (if any) plus whatever has been entered so far, and allow the user
+                // to add arbitrary IRIs via the current input.
+                var choices = new LinkedHashSet<>(getModelObject());
+                choices.addAll(availableDatasets);
+                var result = new ArrayList<>(choices);
+                if (isNotBlank(aInput)) {
+                    result.remove(aInput);
+                    result.add(0, aInput);
+                }
+                return result;
+            }
+        };
+        // The legacy "default dataset" has no special role at query time - it is merged into the
+        // query the same way as the additional datasets. So we present a single combined "Datasets"
+        // field. For backwards compatibility we keep storing the first dataset in the legacy
+        // defaultDatasetIri field and only the remaining ones in the additionalDatasetIris.
+        datasetsField.setModel(LambdaModel.of(this::getDatasets, this::setDatasets));
+        datasetsField.add(this::validateDatasets);
+        add(datasetsField);
+
+        // Enumerating the named graphs requires a connection to the endpoint, which is only
+        // available once the knowledge base has been saved (i.e. registered).
+        var loadDatasetsButton = new LambdaAjaxLink("loadDatasets", this::actionLoadDatasets);
+        loadDatasetsButton.add(LambdaBehavior
+                .visibleWhen(() -> kbModel.getObject().getKb().getRepositoryId() != null));
+        loadDatasetsButton.add(new AttributeModifier("title",
+                new StringResourceModel("kb.iri.datasets.load", this)));
+        add(loadDatasetsButton);
+    }
+
+    @Override
+    public void onEvent(IEvent<?> aEvent)
+    {
+        super.onEvent(aEvent);
+
+        if (aEvent.getPayload() instanceof RemoteRepositoryUrlChangedEvent event) {
+            // The discovered graphs belong to the previous endpoint - drop them so we do not offer
+            // datasets that may not exist on the new endpoint.
+            availableDatasets.clear();
+            event.getTarget().add(datasetsField);
+        }
+    }
+
+    private void actionLoadDatasets(AjaxRequestTarget aTarget)
+    {
+        var kb = kbModel.getObject().getKb();
+
+        try {
+            availableDatasets.clear();
+            availableDatasets.addAll(kbService.listDatasets(kb));
+            success("Loaded [" + availableDatasets.size() + "] graph(s) from the endpoint.");
+        }
+        catch (RuntimeException e) {
+            availableDatasets.clear();
+            error("Unable to load graphs from the endpoint: " + e.getMessage());
+        }
+
+        aTarget.add(datasetsField);
+        aTarget.addChildren(getPage(), IFeedback.class);
+    }
+
+    private Collection<String> getDatasets()
+    {
+        var kb = kbModel.getObject().getKb();
+        var datasets = new LinkedHashSet<String>();
+        if (isNotBlank(kb.getDefaultDatasetIri())) {
+            datasets.add(kb.getDefaultDatasetIri());
+        }
+        datasets.addAll(kb.getAdditionalDatasetIris());
+        return datasets;
+    }
+
+    private void setDatasets(Collection<String> aDatasets)
+    {
+        var kb = kbModel.getObject().getKb();
+        var datasets = aDatasets != null ? new ArrayList<>(new LinkedHashSet<>(aDatasets))
+                : new ArrayList<String>();
+        if (datasets.isEmpty()) {
+            kb.setDefaultDatasetIri(null);
+            kb.setAdditionalDatasetIris(emptyList());
+        }
+        else {
+            kb.setDefaultDatasetIri(datasets.get(0));
+            kb.setAdditionalDatasetIris(datasets.subList(1, datasets.size()));
+        }
+    }
+
+    private void validateDatasets(IValidatable<Collection<String>> aValidatable)
+    {
+        for (var iri : aValidatable.getValue()) {
+            var validatable = new Validatable<String>(iri);
+            Validators.IRI_VALIDATOR.validate(validatable);
+            validatable.getErrors().forEach(aValidatable::error);
+        }
     }
 
     private CheckBox maxQueryLimitCheckbox(String id, IModel<Boolean> aModel)
