+++ inception/inception-ui-kb/src/main/java/de/tudarmstadt/ukp/inception/ui/kb/project/RemoteRepositoryUrlChangedEvent.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed to the Technische Universität Darmstadt under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The Technische Universität Darmstadt
+ * licenses this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.tudarmstadt.ukp.inception.ui.kb.project;
+
+import org.apache.wicket.ajax.AjaxRequestTarget;
+
+/**
+ * Broadcast when the URL of a remote knowledge base is changed. Components that derive choices from
+ * the endpoint (e.g. the datasets field) can listen for this to drop now-stale information.
+ */
+public class RemoteRepositoryUrlChangedEvent
+{
+    private final AjaxRequestTarget target;
+
+    public RemoteRepositoryUrlChangedEvent(AjaxRequestTarget aTarget)
+    {
+        target = aTarget;
+    }
+
+    public AjaxRequestTarget getTarget()
+    {
+        return target;
+    }
+}
