+++ inception/inception-ui-kb/src/main/resources/META-INF/asciidoc/user-guide/projects_knowledge-base.adoc
@@ -70,12 +70,6 @@ The remote knowledge bases, there are the following settings:
   Consult with the operator of the SPARQL server if you are unsure which datasets are available.
   If no dataset is selected, queries run against the default graph.
 
-NOTE: **Changing the URL of a remote KB currently only takes affect after {product-name} is restarted!**
-      The updated URL will be shown in the settings, but queries will still be sent to the old URL until you restart {product-name}.
-      This also means that if you add, remove or change HTTP "Basic" authentication that are part of the URL, they will not
-      take effect until you restart. It is usually easier to delete the remote KB configuration and create it from scratch
-      with the new URL.
-
 === General settings
 
 * **Base prefix:** A base IRI that is used as a prefix when creating new concepts.
